backend/requirements.txt
backtrader==1.9.78.123
certifi==2021.10.8
chardet==3.0.4
Django==5.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==2.10
macrocosmos==0.1.4
matplotlib==3.5.1
numpy==1.22.0
pandas==2.2.3
plotly==6.0.1
python-binance==1.0.15
python-dotenv==0.21.0
requests==2.27.1
seaborn==0.11.2
urllib3==1.26.9
cbpro==1.1.4
